move cost model to cost.py
